refactor(matrix-call-api): extract shared room-call-permission helpers
Move shared auth, Matrix request, token resolution, event constants, and power-level types into a single _lib module and reuse it across repair and diagnostics routes.
